[TASK] Allow to extend the default index behavior via fallback rules
This commit introduces a new option 'fallback' that can be passed to
the ComponentDomParser's constructor. Und that key you're now able to
define a map of matchers and functions.

Whenever a data-* component cannot be found the ComponentDomParser will
now look up a matching fallback rule to let it handle the lookup for a
component.

This allows for implementing a central component lookup even if a component
name is unknown during compile time.